refactor: consolidate grid and top bar components for major code reduction
- Create BaseGrid component to unify grid layouts across MemoryGrid and GalleryGrid
- Create BaseTopBar component to unify top bar layouts across GalleryTopBar, DashboardTopBar, and FolderTopBar
- Refactor all grid components to use BaseGrid (eliminates ~150+ lines)
- Refactor all top bar components to use BaseTopBar (eliminates ~450+ lines)
- Reorganize component structure: move GalleryTopBar to galleries/, DashboardTopBar to dashboard/
- Fix infinite loop bug in BaseTopBar useEffect
- Update all import paths to reflect new structure
- Total: ~600+ lines of duplicate code eliminated
- Improves maintainability, consistency, and follows React best practices

Components affected:
- MemoryGrid, GalleryGrid → BaseGrid
- GalleryTopBar, DashboardTopBar, FolderTopBar → BaseTopBar
